chore: adjust code style according to launch review
